+++ changelog.md
@@ -2,4 +2,5 @@
 
 ## Version 0.1.0 (2023/02/20)
 
+- Added bible book identification, and much more!
 - Init